requirements.txt
blinker==1.9.0
Flask==3.1.1
matplotlib==3.10.3
numpy==2.1.3
pandas==2.3.0
tensorflow_cpu==2.19.0
gunicorn